+++ upload/admin/language/en-gb/extension/ventocart/module/gdpr.php
@@ -1,0 +1,17 @@
+<?php
+// Heading
+$_['heading_title'] = 'GDPR Notice';
+
+// Text
+$_['text_extension'] = 'Extensions';
+$_['text_success'] = 'Success: You have modified GDPR module!';
+$_['text_edit'] = 'Edit Option GDPR module';
+$_['text_notice_message'] = "Notice Message";
+$_['help_force_all'] = 'This will force them to accept both tracking and essential cookies or leave the site.';
+$_['help_gdpr'] = 'If the user denies consent for tracking cookies, the abandoned cart and analytics modules will not function for them.';
+
+// Entry
+$_['entry_status'] = 'Status';
+
+// Error
+$_['error_permission'] = 'Warning: You do not have permission to modify GDPR module!';