+++ upload/catalog/controller/extension/ventocart/analytics/google_analytics.php
@@ -8,18 +8,25 @@ namespace Ventocart\Catalog\Controller\Extension\VentoCart\Analytics;
  */
 class GoogleAnalytics extends \Ventocart\System\Engine\Controller
 {
-	/**
-	 * @return string
-	 */
-	public function index(): string {
-		$myTag = $this->model_setting_setting->getValue('analytics_google_analytics_tag', $this->config->get('config_store_id'));
-	
-        $fixer = md5(rand(1000000,1000000000));
-		// Encode the tag ID using Base64 in PHP
-		$encodedTag = base64_encode( $fixer . $myTag);
-	
-		// Escape the single quotes in the JavaScript string by doubling them
-	    return "<!-- Google tag (gtag.js) -->
+    /**
+     * @return string
+     */
+    public function index(): string
+    {
+
+        if ($this->config->get('module_gdpr_status') && !isset($this->request->cookie['accept-tracking'])) {
+            return '';
+        }
+
+
+        $myTag = $this->model_setting_setting->getValue('analytics_google_analytics_tag', $this->config->get('config_store_id'));
+
+        $fixer = md5(rand(1000000, 1000000000));
+        // Encode the tag ID using Base64 in PHP
+        $encodedTag = base64_encode($fixer . $myTag);
+
+        // Escape the single quotes in the JavaScript string by doubling them
+        return "<!-- Google tag (gtag.js) -->
             <script>
                 // Custom function for decoding
                 function jtagdc(encoded) {
@@ -41,9 +48,9 @@ class GoogleAnalytics extends \Ventocart\System\Engine\Controller
                     gtag('config', jtagdc('$encodedTag')); // Pass the decoded tag ID to gtag
                 };
             </script>";
-	}
- 
+    }
 
- 
-	
+
+
+
 }