+++ upload/catalog/language/en-gb/extension/ventocart/module/gdpr.php
@@ -1,0 +1,7 @@
+<?php
+// Text Disclaimer
+$_['text_notice'] = 'We are using cookies. Please read our %s and %s for more information.';
+$_['button_accept'] = 'Accept cookies';
+$_['button_deny'] = 'Deny';
+$_['text_tracking_cookies'] = 'Tracking Cookies';
+$_['text_essential_cookies'] = 'Essential Cookies (required)';